requirements.txt
opencv-python
numpy
Pillow
faiss-cpu
pytesseract
layoutparser  # opcional; instalación puede requerir Detectron2 (complejo)
#detectron2 -f https://dl.fbaipublicfiles.com/detectron2/wheels/cu113/torch1.10/index.html   instalar cuando tienes solo CPU
#detectron2 -f https://dl.fbaipublicfiles.com/detectron2/wheels/cu113/torch1.10/index.html   instalar cuando tienes GPU
#openai # opcional
easyocr
#realesrgan
#pip install layoutparser torchvision && pip install "detectron2@git+https://github.com/facebookresearch/detectron2.git@v0.5#egg=detectron2"}
#agregar esto a la instalacion
pix2tex[gui]
#librerias para hacer el procesamiento de imagen a texto
transformers
torch
sentence-transformers

#implementacion para rag
langchain
langchain-community chromadb sentence-transformers pypdf pdfp lumber
ollama
llama-cpp-python 
beautifulsoup4